(daoadmin): create sidebar component file
